:bug: Make sure nav bar is visible in every view (#879)
* Use common `NavigationController` for all the view controllers instantiated for tabs presentation

* Improve readability

* Add proper spacing between navigation controller items

* Update `icCalendar` icon

* Show calendar button on navigation controller bar only in `GoalsViewController`

* Move calendar button to `MainViewController`

* Delete `GoalNavigationController` and related stuff, no need in it since we unified `NavigationController` usage

* :speak_no_evil:

* Update tests
