+++ teferi/UI/Modules/Navigation/NavigationController.swift
@@ -9,7 +9,6 @@ class NavigationController: UINavigationController
     private var viewModel : NavigationViewModel!
     private var presenter : NavigationPresenter!
     
-    private var calendarButton : UIButton!
     private var feedbackButton : UIButton!
     private var logoImageView : UIImageView!
     private var titleLabel : UILabel!
@@ -27,16 +26,7 @@ class NavigationController: UINavigationController
     override func viewDidLoad()
     {
         super.viewDidLoad()
-        
-        calendarButton = UIButton(frame: CGRect(x: 0, y: 0, width: 40, height: 40))
-        calendarButton.titleLabel?.font = UIFont.systemFont(ofSize: 11)
-        calendarButton.setBackgroundImage(Image(asset: Asset.icCalendar), for: .normal)
-        calendarButton.rx.tap
-            .subscribe(onNext: { [unowned self] in
-                self.presenter.toggleCalendar()
-            })
-            .addDisposableTo(disposeBag)
-        
+
         feedbackButton = UIButton(frame: CGRect(x: 0, y: 0, width: 40, height: 40))
         feedbackButton.setBackgroundImage(Image(asset: Asset.icFeedback), for: .normal)
         feedbackButton.rx.tap
@@ -72,10 +62,6 @@ class NavigationController: UINavigationController
     
     private func bindViewModel()
     {
-        viewModel.calendarDay
-            .bindTo(calendarButton.rx.title(for: .normal))
-            .addDisposableTo(disposeBag)
-        
         viewModel.title
             .bindTo(titleLabel.rx.text)
             .addDisposableTo(disposeBag)
@@ -85,29 +71,20 @@ class NavigationController: UINavigationController
     {
         super.pushViewController(viewController, animated: animated)
         
-        setupNavigationBar(viewController: viewController)
+        setupNavigationBar(for: viewController)
     }
     
-    private func setupNavigationBar(viewController: UIViewController)
+    private func setupNavigationBar(for viewController: UIViewController)
     {
-        let calendarBarButtonItem = UIBarButtonItem(customView: calendarButton)
-        let feedbackBarButtonItem = UIBarButtonItem(customView: feedbackButton)
-        
         viewController.navigationItem.rightBarButtonItems = [
-            calendarBarButtonItem,
-            feedbackBarButtonItem
+            .createFixedSpace(of: 2),
+            UIBarButtonItem(customView: feedbackButton)
         ]
         
-        let logoMargin = UIBarButtonItem(barButtonSystemItem: .fixedSpace, target: nil, action: nil)
-        logoMargin.width = 17
-        
-        let bigSpacing = UIBarButtonItem(barButtonSystemItem: .fixedSpace, target: nil, action: nil)
-        bigSpacing.width = 18
-        
         viewController.navigationItem.leftBarButtonItems = [
-            logoMargin,
+            .createFixedSpace(of: 17),
             UIBarButtonItem(customView: logoImageView),
-            bigSpacing,
+            .createFixedSpace(of: 18),
             UIBarButtonItem(customView: titleLabel)
         ]
     }
