+++ teferi/UI/Modules/Navigation/NavigationPresenter.swift
@@ -4,57 +4,21 @@ class NavigationPresenter : NSObject
 {
     private weak var viewController : NavigationController!
     private let viewModelLocator : ViewModelLocator
-    
-    private var calendarViewController : CalendarViewController? = nil
 
     private init(viewModelLocator: ViewModelLocator)
     {
         self.viewModelLocator = viewModelLocator
     }
     
-    static func create(with viewModelLocator: ViewModelLocator) -> NavigationController
+    static func create<T: UIViewController>(with viewModelLocator: ViewModelLocator, rootViewController: T) -> NavigationController
     {
         let presenter = NavigationPresenter(viewModelLocator: viewModelLocator)
         
-        let mainViewController = MainPresenter.create(with: viewModelLocator)
-        let viewController = NavigationController(rootViewController: mainViewController)
-        viewController.inject(presenter: presenter, viewModel: viewModelLocator.getNavigationViewModel(forViewController: viewController))
+        let viewController = NavigationController(rootViewController: rootViewController)
+        viewController.inject(presenter: presenter, viewModel: viewModelLocator.getNavigationViewModel(forViewController: rootViewController))
         
         presenter.viewController = viewController
         
         return viewController
-    }
-    
-    func toggleCalendar()
-    {
-        if let _ = calendarViewController {
-            hideCalendar()
-        } else {
-            showCalendar()
-        }
-    }
-    
-    private func showCalendar()
-    {
-        calendarViewController = CalendarPresenter.create(with: viewModelLocator, dismissCallback: didHideCalendar)
-        viewController.topViewController?.addChildViewController(calendarViewController!)        
-        viewController.topViewController?.view.addSubview(calendarViewController!.view)
-        calendarViewController!.didMove(toParentViewController: viewController.topViewController)
-    }
-    
-    private func hideCalendar()
-    {
-        calendarViewController?.hide()
-    }
-    
-    private func didHideCalendar()
-    {
-        guard let calendar = calendarViewController else { return }
-        
-        calendar.willMove(toParentViewController: nil)
-        calendar.view.removeFromSuperview()
-        calendar.removeFromParentViewController()
-        
-        calendarViewController = nil
     }
 }
