+++ teferiTests/ViewModels/MainViewModelTests.swift
@@ -1,4 +1,5 @@
 import RxSwift
+import RxTest
 import XCTest
 import Nimble
 @testable import teferi
@@ -21,6 +22,11 @@ class MainViewModelTests : XCTestCase
     private var selectedDateService : MockSelectedDateService!
     private var trackEventService: MockTrackEventService!
     private var motionService: MockMotionService!
+    
+    private var scheduler = TestScheduler(initialClock:0)
+    private var dateLabelObserver: TestableObserver<String>!
+    private var disposeBag = DisposeBag()
+    
     
     override func setUp()
     {
@@ -52,6 +58,13 @@ class MainViewModelTests : XCTestCase
                                   trackEventService: trackEventService,
                                   motionService: motionService)
         
+        timeService.mockDate = getDate(withDay: 13)
+        dateLabelObserver = scheduler.createObserver(String.self)
+        
+        viewModel.calendarDay
+            .subscribe(dateLabelObserver)
+            .addDisposableTo(disposeBag)
+        
     }
     
     override func tearDown()
@@ -59,6 +72,37 @@ class MainViewModelTests : XCTestCase
         disposable?.dispose()
     }
     
+    //MARK: Calendar Button Tests
+    func testTheCalendarDayAlwaysReturnsTheCurrentDate()
+    {
+        let dateText = dateLabelObserver.events.last!.value.element!
+        
+        expect(dateText).to(equal("13"))
+    }
+    
+    func testTheCalendarDayAlwaysHasTwoPositions()
+    {
+        appLifecycleService.publish(.movedToBackground)
+        timeService.mockDate = getDate(withDay: 1)
+        appLifecycleService.publish(.movedToForeground(withDailyVotingNotificationDate: nil))
+        
+        let dateText = dateLabelObserver.events.last!.value.element!
+        
+        expect(dateText).to(equal("01"))
+    }
+    
+    func testDateLabelChangesIfDateChangesWhileOnBackground()
+    {
+        appLifecycleService.publish(.movedToBackground)
+        timeService.mockDate = getDate(withDay: 14)
+        appLifecycleService.publish(.movedToForeground(withDailyVotingNotificationDate: nil))
+        
+        let dateText = dateLabelObserver.events.last!.value.element!
+        
+        expect(dateText).to(equal("14"))
+    }
+    
+    //MARK: Slots
     func testTheAddNewSlotsMethodAddsANewSlot()
     {
         var didAdd = false
@@ -105,5 +149,15 @@ class MainViewModelTests : XCTestCase
                                                 category: category,
                                                 categoryWasSetByUser: false,
                                                 tryUsingLatestLocation: false)!
+    }
+    
+    private func getDate(withDay day: Int) -> Date
+    {
+        var dateComponents = DateComponents()
+        dateComponents.year = Date().year
+        dateComponents.month = 1
+        dateComponents.day = day
+        
+        return Calendar.current.date(from: dateComponents)!
     }
 }
